SDP-1641 refactor modals to reuse shared components
