+++ src/components/ApiKeyCreateModal/styles.scss
@@ -6,70 +6,14 @@
     margin-bottom: pxToRem(24);
   }
 
+  &__permissionsDivider {
+    border-bottom: pxToRem(1) solid var(--color-gray-30);
+    margin-bottom: pxToRem(24);
+  }
+
   &__form {
     display: flex;
     flex-direction: column;
     gap: pxToRem(16);
-  }
-
-  &__allowedIPs {
-    font-family: var(--font-family-mono);
-  }
-
-  &__permissions {
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__permissionsHeading {
-    margin-bottom: pxToRem(8);
-    color: var(--color-gray-60);
-  }
-
-  &__permissionsError {
-    color: var(--color-red-60);
-    font-size: pxToRem(14);
-    margin-bottom: pxToRem(8);
-  }
-
-  &__permissionsList {
-    border-top: pxToRem(1) solid var(--color-gray-30);
-    border-bottom: pxToRem(1) solid var(--color-gray-30);
-    display: flex;
-    flex-direction: column;
-    gap: pxToRem(12);
-    padding: pxToRem(16) 0; 
-  }
-
-  &__permissionRow {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  &__permissionLabel {
-    font-weight: var(--font-weight-medium);
-    min-width: pxToRem(120);
-    color: var(--color-gray-90);
-
-    &--bold {
-      font-weight: var(--font-weight-semi-bold);
-    }
-  }
-
-  &__permissionSelect {
-    flex-shrink: 0;
-    min-width: pxToRem(150);
-  }
-
-  &__resourcePermissions {
-    display: flex;
-    flex-direction: column;
-    gap: pxToRem(12);
-
-    &--disabled {
-      opacity: 0.5;
-      pointer-events: none;
-    }
   }
 }
