+++ src/components/ApiKeyUpdateModal/styles.scss
@@ -1,109 +1,14 @@
 @use "./styles-utils.scss" as *;
 
 .EditApiKeyModal {
+  &__description {
+    color: var(--color-gray-60);
+    margin-bottom: pxToRem(24);
+  }
+
   &__form {
     display: flex;
     flex-direction: column;
     gap: pxToRem(16);
-  }
-
-  &__allowedIPs {
-    display: flex;
-    flex-direction: column;
-    font-family: var(--font-family-mono);
-  }
-
-  &__label {
-    display: block;
-    margin-bottom: pxToRem(4);
-    font-weight: var(--font-weight-medium);
-    font-size: pxToRem(14);
-    color: var(--color-gray-90);
-  }
-
-  &__textarea {
-    min-height: pxToRem(80);
-    padding: pxToRem(8);
-    border: 1px solid var(--color-gray-30);
-    border-radius: pxToRem(4);
-    font-size: pxToRem(14);
-    font-family: var(--font-family-mono);
-    resize: vertical;
-    box-sizing: border-box;
-    width: 100%;
-
-    &--error {
-      border-color: var(--color-red-60);
-    }
-  }
-
-  &__error {
-    color: var(--color-red-60);
-    font-size: pxToRem(12);
-    margin-top: pxToRem(4);
-  }
-
-  &__note {
-    font-size: pxToRem(12);
-    color: var(--color-gray-60);
-    margin-top: pxToRem(4);
-  }
-
-  &__permissions {
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__permissionsHeading {
-    margin-bottom: pxToRem(8);
-    color: var(--color-gray-60);
-  }
-
-  &__permissionsError {
-    color: var(--color-red-60);
-    font-size: pxToRem(14);
-    margin-bottom: pxToRem(8);
-  }
-
-  &__permissionsList {
-    border-top: pxToRem(1) solid var(--color-gray-30);
-    border-bottom: pxToRem(1) solid var(--color-gray-30);
-    display: flex;
-    flex-direction: column;
-    gap: pxToRem(12);
-    padding: pxToRem(16) 0; 
-  }
-
-
-  &__permissionRow {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  &__permissionLabel {
-    font-weight: var(--font-weight-medium);
-    min-width: pxToRem(120);
-    color: var(--color-gray-90);
-
-    &--bold {
-      font-weight: var(--font-weight-semi-bold);
-    }
-  }
-
-  &__permissionSelect {
-    flex-shrink: 0;
-    min-width: pxToRem(150);
-  }
-
-  &__resourcePermissions {
-    display: flex;
-    flex-direction: column;
-    gap: pxToRem(12);
-
-    &--disabled {
-      opacity: 0.5;
-      pointer-events: none;
-    }
   }
 }
